added a cast to make gcc happy when using a pid_t in an sprintf().

+++ src/condor_daemon_core.V6/daemon_core.C
@@ -3035,7 +3035,7 @@ int DaemonCore::Create_Process(
 		return FALSE;
 	}
 
-	sprintf(inheritbuf,"%lu ",mypid);
+	sprintf(inheritbuf,"%lu ",(unsigned long)mypid);
 
 	strcat(inheritbuf,InfoCommandSinfulString());
